fix: 리뷰 반영 - totalAmount Long 타입 변환 + count 정확도 + nullsLast 정렬

+++ SClass-Api-Backoffice/src/main/kotlin/com/sclass/backoffice/coin/dto/AdminCoinTransactionHistoryResponse.kt
@@ -13,7 +13,7 @@ data class AdminCoinTransactionHistoryResponse(
 ) {
     data class CoinTransactionItem(
         val type: CoinTransactionType,
-        val totalAmount: Int,
+        val totalAmount: Long,
         val referenceId: String?,
         val description: String?,
         val createdAt: LocalDateTime,
